feat: task/009-aluno-get-update

+++ src/main/java/school/sptech/back_end_PI/repository/AlunoRepository.java
@@ -1,8 +1,9 @@
 package school.sptech.back_end_PI.repository;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import school.sptech.back_end_PI.entity.Aluno;
 
-public interface AlunoRepository extends JpaRepository<Aluno,Integer> {
+public interface AlunoRepository extends JpaRepository<Aluno, Long> {
 
     boolean existsAlunoByEmail(String email);
